+++ src/objects/hitdetection/ClickHandler.ts
@@ -2,13 +2,23 @@ import { HitEvent } from "../../interfaces/IHitDetection";
 import { HitEventHandler } from "./HitSprite";
 
 export class ClickHandler {
+  public static readonly CLICK = "click";
+  public static readonly POINTER_DOWN = "pointerdown";
+  public static readonly POINTER_UP = "pointerup";
+
   private _doubleClickInfo?: {
     initialEvent: HitEvent;
     timeout: number;
   };
 
+  private _didReceiveClick = false;
+
+  private _map = new Map<string, boolean>();
+
   private _onClick: HitEventHandler | undefined;
   private _onDoubleClick: HitEventHandler | undefined;
+  private _onPointerDown: HitEventHandler | undefined;
+  private _onPointerUp: HitEventHandler | undefined;
 
   public get onClick() {
     return this._onClick;
@@ -26,19 +36,70 @@ export class ClickHandler {
     this._onDoubleClick = value;
   }
 
-  handleClick(event: HitEvent) {
-    if (this.onClick != null || this.onDoubleClick != null) {
-      event.stopPropagation();
-    }
+  public get onPointerDown() {
+    return this._onPointerDown;
+  }
 
+  public set onPointerDown(value) {
+    this._onPointerDown = value;
+  }
+
+  public get onPointerUp() {
+    return this._onPointerUp;
+  }
+
+  public set onPointerUp(value) {
+    this._onPointerUp = value;
+  }
+
+  handleClick(event: HitEvent) {
     if (this._doubleClickInfo == null) {
-      this.onClick && this.onClick(event);
-      this._startDoubleClick(event);
+      if (this._canHandleEvent(ClickHandler.CLICK)) {
+        this._beginEvent(ClickHandler.CLICK);
+
+        this.onClick && this.onClick(event);
+
+        if (this.onDoubleClick != null) {
+          this._startDoubleClick(event);
+        }
+      }
     } else {
-      this.onDoubleClick &&
-        this.onDoubleClick(this._doubleClickInfo.initialEvent);
-      this._resetDoubleClick();
+      this._performDoubleClick(event);
     }
+  }
+
+  handlePointerDown(event: HitEvent) {
+    if (!this._canHandleEvent(ClickHandler.POINTER_DOWN)) return;
+    this._beginEvent(ClickHandler.POINTER_DOWN);
+
+    this.onPointerDown && this.onPointerDown(event);
+  }
+
+  handlePointerUp(event: HitEvent) {
+    if (!this._canHandleEvent(ClickHandler.POINTER_UP)) return;
+    this._beginEvent(ClickHandler.POINTER_UP);
+
+    this.onPointerUp && this.onPointerUp(event);
+  }
+
+  private _beginEvent(name: string) {
+    this._map.set(name, false);
+    setTimeout(() => {
+      this._map.delete(name);
+    }, 0);
+  }
+
+  private _canHandleEvent(name: string) {
+    return this._map.get(name) ?? true;
+  }
+
+  private _performDoubleClick(event: HitEvent) {
+    if (this._doubleClickInfo == null) return;
+
+    event.stopPropagation();
+    this.onDoubleClick &&
+      this.onDoubleClick(this._doubleClickInfo.initialEvent);
+    this._resetDoubleClick();
   }
 
   private _resetDoubleClick() {
@@ -49,6 +110,8 @@ export class ClickHandler {
   }
 
   private _startDoubleClick(event: HitEvent) {
+    event.stopPropagation();
+
     this._doubleClickInfo = {
       initialEvent: event,
       timeout: window.setTimeout(() => this._resetDoubleClick(), 350),