+++ storybook/stories/Furniture.stories.ts
@@ -47,10 +47,13 @@ export function Default() {
       });
 
       furniture.onClick = action(`Furniture ${i} clicked`);
+      furniture.onDoubleClick = action(`Furniture ${i} Double Clicked`);
       furniture.validDirections.then(action(`Furniture ${i} valid directions`));
 
       room.addRoomObject(furniture);
     }
+
+    room.onTileClick = (position) => console.log(position);
 
     const dice = new FloorFurniture({
       roomX: 1,